# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Telemedicina para Todos API",
        "description": "Documentação da API do projeto Telemedicina para Todos. Endpoints internos (sessão) e preparação para API pública de interoperabilidade.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor da aplicação (base atual)"
        },
        {
            "url": "/api/v1",
            "description": "API pública versionada (futuro – interoperabilidade)"
        }
    ],
    "paths": {
...
        "/api/avatar/upload": {
            "post": {
                "tags": [
                    "Avatar"
                ],
                "summary": "Upload de avatar",
                "description": "Upload de avatar do usuário autenticado",
                "operationId": "dfdf9526337b3953d39e90ec6cde9e3c",
                "responses": {
                    "200": {
                        "description": "Avatar atualizado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Arquivo inválido"
                    }
                }
            }
        },
        "/api/avatar/delete": {
            "delete": {
                "tags": [
                    "Avatar"
                ],
                "summary": "Remover avatar",
                "description": "Deletar avatar do usuário autenticado",
                "operationId": "68ac57864cc92f227d5ce340865bcd40",
                "responses": {
                    "200": {
                        "description": "Avatar removido"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Usuário sem avatar"
                    }
                }
            }
        },
        "/api/avatar/show": {
            "get": {
                "tags": [
                    "Avatar"
                ],
                "summary": "URL do avatar",
                "description": "Obter URL do avatar do usuário autenticado",
                "operationId": "eb0bdfb6e2ec32984b36c97436eefdd2",
                "parameters": [
                    {
                        "name": "thumbnail",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "avatar_url"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/doctors/{doctor}/availability/{date}": {
            "get": {
                "tags": [
                    "Disponibilidade"
                ],
                "summary": "Disponibilidade do médico por data",
                "description": "Retorna slots disponíveis do médico para a data informada. Rota pública.",
                "operationId": "b0a503d0ecf77010bdf479d574b57ebc",
                "parameters": [
                    {
                        "name": "doctor",
                        "in": "path",
                        "description": "ID do médico",
                        "required": true
                    },
                    {
                        "name": "date",
                        "in": "path",
                        "description": "Data (Y-m-d)",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Slots disponíveis"
                    },
                    "422": {
                        "description": "Data inválida"
                    }
                }
            }
        },
...
    },
...
}